Uniform namespace for actors across plugins
Removed ActorComponents from actor namespaces
