+++ src/Pixel.Automation.Core.Components/Applications/CloseApplicationActorComponent.cs
@@ -7,7 +7,7 @@ using System.ComponentModel;
 using System.Runtime.Serialization;
 using System.Threading.Tasks;
 
-namespace Pixel.Automation.UIA.Components.ActorComponents
+namespace Pixel.Automation.UIA.Components
 {
     /// <summary>
     /// Use <see cref="CloseApplicationActorComponent"/> to close application which was previously launched or attached to.
