+++ src/Pixel.Automation.UIA.Components/ActorComponents/InvokeActionComponent.cs
@@ -6,7 +6,7 @@ using System.Runtime.Serialization;
 using System.Threading.Tasks;
 using uiaComWrapper::System.Windows.Automation;
 
-namespace Pixel.Automation.UIA.Components.ActorComponents
+namespace Pixel.Automation.UIA.Components
 {
     /// <summary>
     /// Use <see cref="InvokeActorComponent"/> to perform defult stateless action on a control e.g. click on a button.
