+++ src/Pixel.Automation.UIA.Components/ActorComponents/MoveActorComponent.cs
@@ -9,7 +9,7 @@ using System.Runtime.Serialization;
 using System.Threading.Tasks;
 using uiaComWrapper::System.Windows.Automation;
 
-namespace Pixel.Automation.UIA.Components.ActorComponents
+namespace Pixel.Automation.UIA.Components
 {
     /// <summary>
     /// Use <see cref="MoveActorComponent"/> to move a control to new position.
