+++ src/Pixel.Automation.UIA.Components/ActorComponents/ScrollViewActorComponent.cs
@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 using uiaComWrapper::System.Windows.Automation;
 
 
-namespace Pixel.Automation.UIA.Components.ActorComponents
+namespace Pixel.Automation.UIA.Components
 {
     /// <summary>
     /// Use <see cref="ScrollViewActorComponent"/> to perform a horizontal and/or vertical scroll.
