+++ src/Pixel.Automation.UIA.Components/ActorComponents/SelectControlActorComponent.cs
@@ -8,7 +8,7 @@ using System.Runtime.Serialization;
 using System.Threading.Tasks;
 using uiaComWrapper::System.Windows.Automation;
 
-namespace Pixel.Automation.UIA.Components.ActorComponents
+namespace Pixel.Automation.UIA.Components
 {
     /// <summary>
     /// Use <see cref="AddToSelectionActorComponent"/> to select a control.
