+++ src/Pixel.Automation.UIA.Components/ActorComponents/SelectListItemActorComponent.cs
@@ -11,7 +11,7 @@ using System.Runtime.Serialization;
 using System.Threading.Tasks;
 using uiaComWrapper::System.Windows.Automation;
 
-namespace Pixel.Automation.UIA.Components.ActorComponents
+namespace Pixel.Automation.UIA.Components
 {
 
     /// <summary>
