+++ src/Pixel.Automation.UIA.Components/ActorComponents/UIAActorComponent.cs
@@ -10,7 +10,7 @@ using System.Runtime.Serialization;
 using System.Threading.Tasks;
 using uiaComWrapper::System.Windows.Automation;
 
-namespace Pixel.Automation.UIA.Components.ActorComponents
+namespace Pixel.Automation.UIA.Components
 {
 
     /// <summary>
