--- src/Unit.Tests/Pixel.Automation.Scripting.Components.Tests/ActorComponents/ExecuteInlineScriptActorComponentFixture.cs
@@ -1,31 +1,0 @@
-﻿using NSubstitute;
-using NUnit.Framework;
-using Pixel.Automation.Core;
-using Pixel.Automation.Core.Interfaces;
-using Pixel.Automation.Core.Models;
-using System.Threading.Tasks;
-
-namespace Pixel.Automation.Scripting.Components.Tests.ActorComponents
-{
-    public class ExecuteInlineScriptActorComponentFixture
-    {
-        [Test]
-        public async Task CanAct()
-        {
-            var entityManager = Substitute.For<IEntityManager>();
-
-            IScriptEngine scriptEngine = Substitute.For<IScriptEngine>();
-            scriptEngine.ExecuteFileAsync(Arg.Any<string>()).Returns(new ScriptResult());
-            entityManager.GetScriptEngine().Returns(scriptEngine);
-            var scriptedAssignActor = new ExecuteInlineScriptActorComponent()
-            {
-                EntityManager = entityManager,
-                ScriptFile = "script.csx"
-            };
-
-            await scriptedAssignActor.ActAsync();
-
-            await scriptEngine.Received(1).ExecuteFileAsync("script.csx");
-        }
-    }
-}
